Address review feedback on result patches and refs
