+++ AGENTS.md
@@ -53,7 +53,7 @@ Release-oriented notes also live in [`docs/RELEASE.md`](docs/RELEASE.md) under p
 
 `agent_browser` results include stable machine-readable fields on `details`: `resultCategory` (`success` | `failure`), plus `successCategory` or `failureCategory` with small fixed enums. The human-facing contract and field list live in [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#details).
 
-- **Pi transcript alignment:** Pi’s custom-tool path normally marks a tool row failed only when `execute` throws, so a wrapper that returns success while setting `details.resultCategory: "failure"` (for example after `qa` diagnostic reclassification) would otherwise look like a passing tool in JSONL/UI. The extension registers `pi.on("tool_result", …)` and `buildAgentBrowserToolResultPatch` in `extensions/agent-browser/index.ts` patches matching `agent_browser` results to set `isError: true` and append the visible `Result category: failure; …; Pi tool isError: true.` line when `event.isError` was not already true. When changing that behavior, update [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#details), [`docs/ARCHITECTURE.md`](docs/ARCHITECTURE.md), release notes if transcript review guidance shifts, and extension-validation expectations that assert the notice text.
+- **Pi transcript alignment:** Pi’s custom-tool path normally marks a tool row failed only when `execute` throws, so a wrapper that returns success while setting `details.resultCategory: "failure"` (for example after `qa` diagnostic reclassification) would otherwise look like a passing tool in JSONL/UI. The extension registers `pi.on("tool_result", …)` and `buildAgentBrowserToolResultPatch` in `extensions/agent-browser/index.ts` patches matching `agent_browser` results to set `isError: true` when `event.isError` was not already true. For prose results it also appends the visible `Result category: failure; …; Pi tool isError: true.` line; for caller-requested `--json` results it must preserve the parseable JSON string and patch only `isError`. When changing that behavior, update [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#details), [`docs/ARCHITECTURE.md`](docs/ARCHITECTURE.md), release notes if transcript review guidance shifts, and extension-validation expectations that assert both prose notice text and JSON preservation.
 - **Source of truth for enums and classifiers:** `extensions/agent-browser/lib/results/shared.ts` (`classifyAgentBrowserSuccessCategory`, `classifyAgentBrowserFailureCategory`, `buildAgentBrowserResultCategoryDetails`). The tool entrypoint merges category details in `extensions/agent-browser/index.ts`; presentation-layer failures also attach categories from `extensions/agent-browser/lib/results/presentation.ts`.
 - **Structured follow-ups:** optional `details.nextActions` (and per-step `batchSteps[].nextActions`) are built by `buildAgentBrowserNextActions` in the same `shared.ts` module, then merged into Pi-facing `details` from `extensions/agent-browser/index.ts` and `extensions/agent-browser/lib/results/presentation.ts`. See [`docs/TOOL_CONTRACT.md`](docs/TOOL_CONTRACT.md#details) for the field contract.
 - **Page-change summaries:** optional `details.pageChangeSummary` (and per-step `batchSteps[].pageChangeSummary` when a step carries one) compress navigation, mutation, confirmation, and artifact signals into a small struct with `changeType`, `summary`, optional `title`/`url`, artifact fields, and `nextActionIds` that line up with `nextActions`. Types live in `extensions/agent-browser/lib/results/shared.ts` (`AgentBrowserPageChangeSummary`); assembly and the mutation-prone command allowlist are in `extensions/agent-browser/lib/results/presentation.ts` (`buildPageChangeSummary`, `PAGE_CHANGE_SUMMARY_COMMANDS`); successful paths merge from `extensions/agent-browser/index.ts` alongside presentation output.
